requirements.txt
notebook
numpy
matplotlib
scikit-image
pathlib
tqdm
jupyter
pyyaml